CONTINUING TO REFACTOR - move and rename everything
